backend-main/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.9.0
gunicorn==23.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
psycopg==3.3.2
Pillow==12.1.0
cryptography==46.0.4
black==26.1.0
pytest==9.0.2
pytest-django==4.11.1
pytest-cov==7.0.0
psycopg2-binary==2.9.11